M02: slim utils.py; quarantine v2 helpers; replace hpu calls in network.py
- hpvsim/utils.py: slimmed from 785 LOC to 24 LOC; all v2 helpers removed
  since active code no longer imports from utils
- hpvsim/_v2_legacy/utils.py: new file; verbatim copy of pre-M02 utils
  surface so _v2_legacy modules (from . import utils as hpu) continue to
  resolve without touching active code
- hpvsim/network.py: replaced three hpu.* calls with starsim-native dists
  (ss.bernoulli for binomial_filter, ss.choice for choose_w, inlined
  participation_filter as _participation_filter method); dropped hpu import
  entirely; added _dist_cross_f, _dist_cross_m, _dist_participate,
  _dist_choose_m as CRN-safe bernoulli/choice streams
